cpu: aarch64: restore reorder backport and fix SVE guard
