+++ src/common/memory_desc_wrapper.hpp
@@ -33,8 +33,10 @@ namespace dnnl {
 namespace impl {
 
 status_t fill_blocked(memory_desc_t &md, std::vector<int> &perm,
-        std::vector<int> &inner_blks,
-        std::vector<int> &inner_idxs);
+        std::vector<int> &inner_blks, std::vector<int> &inner_idxs);
+
+status_t fill_blocked(memory_desc_t &md, std::vector<dim_t> &perm,
+        std::vector<dim_t> &inner_blks, std::vector<dim_t> &inner_idxs);
 
 /** thin wrapper class over \struct memory_desc_t which allows easy
  * manipulations with underlying C structure, which is taken by reference */
@@ -323,13 +325,20 @@ struct memory_desc_wrapper : public c_compatible {
                 const size_t metadata = padded_dims()[0] * padded_dims()[1] / 64
                         * sizeof(uint64_t);
                 using comp_tile_len_type = int;
-                size_t comp_tile_data_size = ceil(static_cast<float>(padded_dims()[0] * padded_dims()[1])
-                        / (64 * 64 * (64 / sizeof(comp_tile_len_type)))) * 64;
-                return comp_tile_data_size + (padded_dims()[0] * padded_dims()[1] * data_type_size())
+                size_t comp_tile_data_size
+                        = ceil(static_cast<float>(
+                                       padded_dims()[0] * padded_dims()[1])
+                                  / (64 * 64
+                                          * (64 / sizeof(comp_tile_len_type))))
+                        * 64;
+                return comp_tile_data_size
+                        + (padded_dims()[0] * padded_dims()[1]
+                                * data_type_size())
                         + metadata + 1000;
-                        // todo: [av] why 1000?
+                // todo: [av] why 1000?
             } else {
-                printf("encoding:%d\n", (int)sparse_desc().encoding), fflush(stdout);
+                printf("encoding:%d\n", (int)sparse_desc().encoding),
+                        fflush(stdout);
                 assert(!"unknown sparse encoding");
                 return 0;
             }
@@ -446,7 +455,8 @@ struct memory_desc_wrapper : public c_compatible {
         if (utils::one_of(format_kind(), format_kind::undef, format_kind::any))
             return false;
         if (has_runtime_dims_or_strides() || has_broadcast()) return false;
-        return utils::div_up(nelems(with_padding)* data_type_size(), sub_byte_data_type_multiplier())
+        return utils::div_up(nelems(with_padding) * data_type_size(),
+                       sub_byte_data_type_multiplier())
                 == size(0, /* include_additional_size = */ false);
     }
 
@@ -530,8 +540,9 @@ struct memory_desc_wrapper : public c_compatible {
      * following statement might be true: lhs == rhs && !lhs.similar_to(rhs) */
     /* TODO: revise */
     bool similar_to(const memory_desc_wrapper &rhs, bool with_padding = true,
-            bool with_data_type = true, int dim_start = 0, bool use_weak_cmp = false,
-            bool check_off0 = false, uint64_t stride_mask = 0xffffffffffffffff) const;
+            bool with_data_type = true, int dim_start = 0,
+            bool use_weak_cmp = false, bool check_off0 = false,
+            uint64_t stride_mask = 0xffffffffffffffff) const;
 
     /** returns true if one memory can be reordered to another */
     bool consistent_with(const memory_desc_wrapper &rhs) const;
@@ -670,9 +681,8 @@ struct memory_desc_wrapper : public c_compatible {
             memory_desc_t &memory_desc, format_tag_t tag);
 
     static status_t compute_blocking(format_tag_t tag,
-                                     std::vector<size_t> &perm,
-                                     std::vector<size_t> &inner_blks,
-                                     std::vector<size_t> &inner_idxs);
+            std::vector<size_t> &perm, std::vector<size_t> &inner_blks,
+            std::vector<size_t> &inner_idxs);
 
 private:
     /* TODO: put logical_offset in utils */
@@ -703,7 +713,8 @@ private:
 };
 
 inline bool memory_desc_wrapper::similar_to(const memory_desc_wrapper &rhs,
-        bool with_padding, bool with_data_type, int dim_start, bool use_weak_cmp, bool check_off0, uint64_t stride_mask) const {
+        bool with_padding, bool with_data_type, int dim_start,
+        bool use_weak_cmp, bool check_off0, uint64_t stride_mask) const {
     using namespace utils;
 
     if (one_of(format_kind(), format_kind::undef, format_kind::any))
@@ -718,12 +729,16 @@ inline bool memory_desc_wrapper::similar_to(const memory_desc_wrapper &rhs,
     auto custom_cpm = use_weak_cmp ? array_cmp_weak : array_cmp<dnnl_dim_t>;
     auto cmp_strides = [&]() {
         if (0xffffffffffffffff == stride_mask) {
-            return custom_cpm(blk.strides + ds, r_blk.strides + ds, ndims() - ds);
+            return custom_cpm(
+                    blk.strides + ds, r_blk.strides + ds, ndims() - ds);
         } else {
             for (int i = 0; i < ndims(); ++i) {
                 if (stride_mask & (1 << i)) {
                     if (blk.strides[i] != r_blk.strides[i]
-                        && IMPLICATION(use_weak_cmp, (blk.strides[i] != DNNL_RUNTIME_DIM_VAL && r_blk.strides[i] != DNNL_RUNTIME_DIM_VAL))) {
+                            && IMPLICATION(use_weak_cmp,
+                                    (blk.strides[i] != DNNL_RUNTIME_DIM_VAL
+                                            && r_blk.strides[i]
+                                                    != DNNL_RUNTIME_DIM_VAL))) {
                         return false;
                     }
                 }
@@ -736,8 +751,7 @@ inline bool memory_desc_wrapper::similar_to(const memory_desc_wrapper &rhs,
             && format_kind() == rhs.format_kind()
             && IMPLICATION(with_data_type, data_type() == rhs.data_type())
             && custom_cpm(dims() + ds, rhs.dims() + ds, ndims() - ds)
-            && cmp_strides()
-            && blk.inner_nblks == r_blk.inner_nblks
+            && cmp_strides() && blk.inner_nblks == r_blk.inner_nblks
             && array_cmp(blk.inner_blks, r_blk.inner_blks, blk.inner_nblks)
             && array_cmp(blk.inner_idxs, r_blk.inner_idxs, blk.inner_nblks)
             && IMPLICATION(with_padding,
@@ -746,7 +760,10 @@ inline bool memory_desc_wrapper::similar_to(const memory_desc_wrapper &rhs,
                                     rhs.padded_dims() + ds, ndims() - ds)
                             && custom_cpm(padded_offsets() + ds,
                                     rhs.padded_offsets() + ds, ndims() - ds))
-            && IMPLICATION(check_off0, (offset0() == DNNL_RUNTIME_DIM_VAL || rhs.offset0() ==DNNL_RUNTIME_DIM_VAL || offset0() == rhs.offset0()));
+            && IMPLICATION(check_off0,
+                    (offset0() == DNNL_RUNTIME_DIM_VAL
+                            || rhs.offset0() == DNNL_RUNTIME_DIM_VAL
+                            || offset0() == rhs.offset0()));
 }
 
 inline bool memory_desc_wrapper::consistent_with(
