+++ src/cpu/reorder/cpu_reorder.hpp
@@ -37,10 +37,11 @@
 #include "cpu/x64/jit_uni_reorder_direct_copy.hpp"
 #include "cpu/x64/matmul/brgemm_matmul_reorders.hpp"
 #elif DNNL_AARCH64
-#include "cpu/aarch64/jit_uni_reorder.hpp"
 #include "cpu/aarch64/matmul/brgemm_matmul_reorders.hpp"
+#include "cpu/aarch64/reorder/jit_blk_reorder.hpp"
+#include "cpu/aarch64/reorder/jit_uni_reorder.hpp"
 #if defined(DNNL_AARCH64_USE_ACL)
-#include "cpu/aarch64/acl_reorder.hpp"
+#include "cpu/aarch64/reorder/acl_reorder.hpp"
 #endif
 #elif DNNL_PPC64
 #include "cpu/ppc64/ppc64_gemm_reorder.hpp"
